Wrap example require_once in is_readable() check
If the plugin is being loaded as a dependency of something larger (e.g.
the project is loading plugins via Composer), the vendor directory may
not be local to this plugin. We don't want to throw a fatal error if
this is the case, so we check the file exists and is readable first.

Fixes #28