Add intermediary mechanism for processing shortcode data.
The previous iteration of the Shortcode class made it possible to create an empty shortcode, but not
one which processed attributes or content included in the shortcode. This commit introduces two new
protected class properties - attributes and content - which can be accessed by the extending class
to perform any data sanitization or value checking prior to rendering.

By default, the Shortcode abstract will simply set those values on the class, then call its render
method, which the extending class is still required to implement. This provides the benefit of
keeping the render interface the same (preserving backward compatibility) while also enabling
engineers to work with shortcode data in a way similar to how they would expect.